First pass of Message integration
I stopped progress on this because I was asked to do something else and
also because of questions on how to integrate it with the Inventory
BoltOn. These changes add Message information to an entirely new page
but we concluded it'd be best to intelligently combine Message and
Inventory displays into a single central table if they were both
enabled (and only display Inventory information if not).
Ideally, this central table would reconcile the case in which an asset
only has Inventory information and not Message or vice versa and would
display an asset's status if it was present.

Also, since the writing of this, I've changed the output of `flight
message show` and as such the controller won't be entirely functional
from the get-go, you'll need to tinker with the regex a little.

Also the CSS is really bad.

I hope this can be of use to someone.